Use double

+++ src/py/migraphx_py.cpp
@@ -547,7 +547,7 @@ MIGRAPHX_PYBIND11_MODULE(migraphx, m)
         py::arg("format") = "msgpack");
 
     m.def("get_target", &migraphx::make_target);
-    m.def("create_argument", [](const migraphx::shape& s, const std::vector<float>& values) {
+    m.def("create_argument", [](const migraphx::shape& s, const std::vector<double>& values) {
         if(values.size() != s.elements())
             MIGRAPHX_THROW("Values and shape elements do not match");
         migraphx::argument a{s};
